requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain_groq
sentence-transformers
pinecone-client
pydantic
groq